feat(player, agents): merge `players` with `agents`
