+++ src/plugins/agents/Cargo.toml
@@ -12,7 +12,12 @@ serde.workspace = true
 # internal
 common.workspace = true
 macros.workspace = true
+zyheeda_core.workspace = true
 
 [dev-dependencies]
+# external
+mockall.workspace = true
+test-case.workspace = true
+
 # internal
 testing.workspace = true
