+++ src/plugins/agents/src/components.rs
@@ -1,1 +1,5 @@
-pub mod enemy;
+pub(crate) mod enemy;
+pub(crate) mod player;
+pub(crate) mod player_camera;
+pub(crate) mod player_movement;
+pub(crate) mod skill_animation;
