+++ src/plugins/agents/src/components/player.rs
