+++ src/plugins/agents/src/components/player_camera.rs
