+++ src/plugins/agents/src/components/skill_animation.rs
