+++ src/plugins/agents/src/resources.rs
