+++ src/plugins/agents/src/resources/cam_ray.rs
