+++ src/plugins/agents/src/resources/mouse_hover.rs
