+++ src/plugins/agents/src/systems.rs
@@ -1,1 +1,4 @@
+pub(crate) mod set_cam_ray;
+pub(crate) mod set_mouse_hover;
+pub(crate) mod toggle_walk_run;
 pub(crate) mod void_sphere;
