+++ src/plugins/agents/src/systems/set_cam_ray.rs
