+++ src/plugins/agents/src/systems/set_mouse_hover.rs
