--- src/plugins/player/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "player"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-# external
-bevy.workspace = true
-bevy_rapier3d.workspace = true
-serde.workspace = true
-
-# internal
-common.workspace = true
-macros.workspace = true
-zyheeda_core.workspace = true
-
-[dev-dependencies]
-# external
-mockall.workspace = true
-test-case.workspace = true
-
-# internal
-testing.workspace = true
