--- src/plugins/player/src/components.rs
@@ -1,4 +1,0 @@
-pub mod player;
-pub mod player_camera;
-pub mod player_movement;
-pub mod skill_animation;
